+++ lib/src/result_notifier/action_listenable_builder.dart
@@ -5,31 +5,31 @@ import 'package:stated_result/stated_result.dart';
 import 'package:stated_result/stated_result_builder.dart';
 
 class ActionListenableBuilder extends StatelessWidget {
-  final ValueListenable<AsyncActionResult> valueListenable;
+  final ValueListenable<AsyncActionResult> listenable;
   final WidgetBuilder? pendingBuilder;
   final WidgetBuilder? waitingBuilder;
   final FailedResultBuilder? failedBuilder;
-  final WidgetBuilder completedBuilder;
+  final WidgetBuilder builder;
 
   const ActionListenableBuilder({
     Key? key,
     this.pendingBuilder,
     this.waitingBuilder,
     this.failedBuilder,
-    required this.completedBuilder,
-    required this.valueListenable,
+    required this.builder,
+    required this.listenable,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) =>
       ValueListenableBuilder<AsyncActionResult>(
-        valueListenable: valueListenable,
+        valueListenable: listenable,
         builder: (context, result, _) => ActionResultBuilder(
           result: result,
           pendingBuilder: pendingBuilder,
           waitingBuilder: waitingBuilder,
           failedBuilder: failedBuilder,
-          completedBuilder: completedBuilder,
+          builder: builder,
         ),
       );
 }