+++ lib/src/result_notifier/query_listenable_builder.dart
@@ -5,31 +5,31 @@ import 'package:stated_result/stated_result.dart';
 import 'package:stated_result/stated_result_builder.dart';
 
 class QueryListenableBuilder<T> extends StatelessWidget {
-  final ValueListenable<AsyncQueryResult<T>> valueListenable;
+  final ValueListenable<AsyncQueryResult<T>> listenable;
   final WidgetBuilder? pendingBuilder;
   final WidgetBuilder? waitingBuilder;
   final FailedResultBuilder? failedBuilder;
-  final ValueResultBuilder<T> valueBuilder;
+  final ValueResultBuilder<T> builder;
 
   const QueryListenableBuilder({
     Key? key,
     this.pendingBuilder,
     this.waitingBuilder,
     this.failedBuilder,
-    required this.valueBuilder,
-    required this.valueListenable,
+    required this.builder,
+    required this.listenable,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) =>
       ValueListenableBuilder<AsyncQueryResult<T>>(
-        valueListenable: valueListenable,
+        valueListenable: listenable,
         builder: (context, result, _) => QueryResultBuilder<T>(
           result: result,
           pendingBuilder: pendingBuilder,
           waitingBuilder: waitingBuilder,
           failedBuilder: failedBuilder,
-          valueBuilder: valueBuilder,
+          builder: builder,
         ),
       );
 }